Use rotary embeddings from Olmo3

+++ src/transformers/models/olmoe2/modeling_olmoe2.py
@@ -390,8 +390,7 @@ class Olmoe2RotaryEmbedding(nn.Module):
             emb = torch.cat((freqs, freqs), dim=-1)
             cos = emb.cos() * self.attention_scaling
             sin = emb.sin() * self.attention_scaling
-
-        return cos.to(dtype=x.dtype), sin.to(dtype=x.dtype)
+            return cos, sin
 
 
 @auto_docstring
